Move basepath filter
